### 2022-11-08 #4683 @georgiastuart
@JakobBD can you describe the role of the other four authors on the paper? It seems like, from the repository, you're the only committer.
