### 2022-11-14 #4683 @JakobBD
@georgiastuart thank you for reviewing this paper!

> @JakobBD can you describe the role of the other four authors on the paper?

As you already stated, the early stage coding of PoUnce was a joint effort between Thomas, Fabian and myself, as can also be seen from their commits. Andrea Beck and Claus-Dieter Munz contributed along the whole code development process with advice on required code functionality, numerous major and minor code design choices, and reviewing/testing code documentation. This is why I saw the JOSS authorship criterion of “active project direction and other forms of non-code contributions” fulfilled.

> Is this referring to a specific widely-used software for UQ? It definitely isn't inherent in UQ code designs.

Large numbers of model evaluations are inherent to Monte Carlo simulations due to their half-order convergence, especially to Multilevel and Multifidelity approaches, where O(1e5-1e7) evaluations of the cheapest model are common.

If the baseline model is invoked as a standalone tool with its own file I/O, the number of created files is of course at least the number of model evaluations. Most HPC cluster file systems are designed for large files, but not for such large numbers of files. In an earlier approach (prior to PoUnce), we encountered severe performance bottlenecks due to this.

To elaborate a little further:

A common “integrated” alternative is to convert an existing simulation code into a library with a subroutine interface, which is then called from the UQ framework and returns the result, but does not write its own files. However, practical simulation setups often consist of a tool chain of several pre- and post-processing steps, which are often linked through file I/O and require different numbers of processors. This makes converting such a tool chain to a single subroutine without file I/O very cumbersome, if not infeasible. Another disadvantage of this approach is that intermediate steps of the tool chain are lost as they are not saved to a file.

(On a side note, both of these interface types are implemented in the leading UQ software Dakota, and the arguments for and against each of them are pointed out in the [documentation](https://dakota.sandia.gov/sites/default/files/docs/6.16.0/Users-6.16.0.pdf).)

Both approaches are also implemented in PoUnce, but a third (preferred) option is also offered, where we hope to combine the advantages of both approaches. Here, models are executed in a separate process for each step in the tool chain with file I/O, but batches of model evaluations are grouped together as a single process with a common file I/O. This also requires some implementation, but it retains the flexibility and traceability of the standalone approach and the performance of the integrated approach.

I hope this answers your question? I propose to clarify this in the paper by adding

> ... This includes I/O bottlenecks due to very large numbers of relatively small files **(if each baseline model evaluation has its individual file I/O)**, as well as ...
