### 2022-11-19 ce69e44
added possibility to plot histograms as pngs for Monte Carlo internal QoIs
