### 2022-11-25 45db5b9
add documentation on regression tests and random variables
