### 2022-11-25 #4683 @JakobBD
@salrm thanks to you as well for reviewing this paper!

> * After installing the requirements, I could run the tests in `ini/internal_local`. It could be useful to have a short description of the tests in the documentation.

Yes, that makes sense. I’ve added a section `Tests` in the `Getting Started` chapter.

> * It might be helpful to add a short explanation in the paper/documentation about the type of distributions of random variables which can be handled by the library.

Oh yes, thanks for the comment. I’ve added a section to the `Features` chapter of the documentation. I haven’t added it to the paper, since it’s supposed to be very short and, for example, also does not mention the included baseline solvers.

> * Would there be any problem if the simulations at the same level or fidelity do not progress equally in time? I am asking this to see if the framework can be used for simulation of time-dependent physical problems with the possibility of computing stochastic moments on the fly.

Short answer: In the current implementation, there is no problem, apart from potential performance losses for vastly different time steps. Whether and how this works depends on the parallel execution strategy and the implementation of the adaptation of the baseline solver. In our current implementation, the time step is set to the minimal one across parallel samples.

Long answer: There are currently two methods for parallel execution of samples (one based on MPI and one on GNU parallel). In our MPI batch implementation (the one which is replicated in the demonstator_batch solver), PoUnce only does the following: It hands a list of stochastic parameters to the solver along with the number of parallel and sequential runs in a single HDF5 file, it writes the scheduler batch job file (including run command, number of nodes and run time) and collects the results also in a single HDF5 file.

It is agnostic to the rest, which thus depends of the implementation of the baseline solver adaptation.

In our case (with the adaptation of FLEXI), we chose to perform a synchronised output of all parallel runs at different time intervals to one HDF5 file. We therefore set the time step to the minimum across all parallel samples (using MPI_ALLREDUCE). (This also means that more time steps will be computed in some samples than needed for stability reasons, which did not impair performance significantly in our applications). In this ‘MPI/HDF5 batch’ case, several statrategies are possible to compute stochastic moments on the fly:

* The computation of the moments could be carried out in the solver and not in PoUnce. On the PoUnce side, the only required change would be to force-set the number of sequential runs to one, to have all samples available at the same time.
* PoUnce could be adapted to monitor and evaluate files on the fly written by the solver. I don’t see any fundamental obstacles to this, but PoUnce would have to be modified, of course. If this evaluation were used, however, for some kind of on-the-fly solver adaptivity, that would require substantial changes to the code.
* One could also break up the computation into several time sub-intervals, which would then be run as different stages in PoUnce. This would allow to transfer data back to PoUnce and evaluate the moments there after each time sub-interval.

A word on performance: The performance loss due to synchronized time steps is of course determined by how much the time step sizes vary across samples. The problem is inherent in the design choice of synchronous HDF5 I/O. With synchronous I/O, the problem could be somewhat alleviated if the time step variations are predictable and can be estimated from the stochastic input (e.g. due to an uncertain grid spacing): Sorting the samples according to their expected time steps would place samples with similar time steps in parallel, while ones with largely different time steps would be placed in an earlier or later sequential run.

With individual or asynchronous I/O, the baseline solver adaptation could be implemented with an internal scheduler assigning new sample simulations to free compute nodes every time a sample evaluation is finished. But this would not allow to compute stochastic moments on the fly.

> * I have a question about this sentence "Stochastic post-processing can be external (for computationally expensive QoIs such as random fields) or framework-internal". at the bottom of p. 10 of the compiled documentation: does this mean the computational of higher-order stochastic moments of QoIs and/or sample distribution of QoI?

Yes, for example.

Currently, PoUnce only computes mean and standard deviation, and all the external tools we have written do the same.

Regarding the **framework-internal** case: Adding computation of further quantities in PoUnce is quite simple. To demonstrate this, for the Monte Carlo methods, I’ve added the functionality to plot histograms in the branch `stochOutput`, which I’ve added to the GitHub repo. Similarly, since our PCE implementation is based on the ChaosPy library, which also provides computation of higher-order stochastic moments, the according functions can be extended easily.

If **external tools** are used for stochastic post-processing, the only output that is fed back to PoUnce are quantities needed for adaptive sample number computation (such as level difference variances in MLMC). All the rest, including final results, are computed outside of PoUnce, so anything can be added there.
