### 2022-12-01 #4683 @Nikoleta-v3
Thank you very much for your time @salrm8 😄 would you mind generating your checklist with the command:

and go over it? I am sure you haven't missed any of the points but just as a sanity check.

I agree with the reviewer's point about a contribution list/contributions guidelines 👍🏻 They can be part of the docs or you can have a standalone file on the root of the repository.  There you should include a few sentences about running the test suite (I can see that you have a test file).
