### 2022-12-04 #4683 @JakobBD
@salrm8 and @Nikoleta-v3 thank you for the suggestion. I added a [file](https://github.com/JakobBD/pounce/blob/master/CONTRIBUTING.md) to the root of the repository with a list of possible contributions.

I referenced the code documentation, which includes sections on how to extend the code and, upon @salrm8 's sensible suggestion, a description of the tests.
