+++ CONTRIBUTING.md
@@ -30,6 +30,6 @@ Here are some possible ways to extend or improve PoUnce
 ### Refactoring and code quality: 
 
 - Report what is hard to understand when reading into the source code
-- Flatten some of the inheritance schemes, for example by integrating the `Simulation` class in `UqMethod`.
-- Break up some long routines, such as in [src/uqmethod/mfmc.py](uqmethod/mfmc.py) or [src/machines/hawk.py](machines/hawk.py).
+- Flatten some of the inheritance schemes, for example by integrating the [Simulation](src/simulation/simulation.py) class into [UqMethod](src/uqmethod/uqmethod.py).
+- Break up some long routines, such as in [mfmc.py](src/uqmethod/mfmc.py) or [hawk.py](src/machines/hawk.py).
 - Enable linting and maybe add to integration actions
