### 2023-01-04 #4683 @Nikoleta-v3
Hello @JakobBD, and thank you for your patience with your submission.  I have emailed your second reviewers and I am waiting for their reply! In the meantime I had a chance to go over your submission.

Some minor comments:

- In the ReadMe, in the `Quick start / basic run` section there is a typo.  It should be `python3 ../../src/pounce.py parameter_mlmc.yml`.

- Could you also include a sentence about the expected outcome? Tell the user that the results are printed in the terminal and also an output csv is exported.

- In the `Documentation` section could you please include a link to the documentation?

- In the manuscript line 55: it is meant to be interval or internal?

Other issues:

1. In the documentation the sections “Parameters” and “Features” should include more information. In the parameters you should walk the user over all the sections of a `yml` file (`general`, `uq_method`, …).  In the features you should include code snippets with examples of how different methods, solvers etc can be used.
The example `yml` files are great as examples but this information should be more accessible/obvious to a user.

2. To me pounce it’s the perfect project to be a command line tool. So I could just run `pounce parameter_mfmc.yml`. Is there a reason behind your decision to keep it in scripts format? 😊

3. You say that:

> Output files will be located in the current working directory, so it is best to copy a parameter file to an empty directory and start pounce from there.

Why not have the location and the name (if not given by the user) of the output file as arguments?
