### 2023-01-15 f8342ee
add Makefile with mdline target, which converts pounce to a command line tool
