+++ doc/100_getting_started.md
@@ -10,7 +10,7 @@ For a test run, go to `ini/internal_local` and run
 python3 ../../src/pounce.py parameter_mlmc.yml
 ```
 
-## Parameters
+## Parameters (Overview)
 
 The basic syntax for a PoUnce run is
 ```
@@ -35,12 +35,12 @@ Each parameter file contains the sections given in the following table. Most sec
 | `general`         | General settings like the project name and whether or not to save intermediate states of PoUnce. |
 | `uq_method`       | UQ method to be used and general parameters about its configuration (such as the computational budget). | 
 | `sampling`        | Configuration of the sampling method, such as polynomial degree for polynomial chaos expansion or fixed seed of the random number generator for Monte Carlo. |
-| `solver`/`models` | Setup of the baseline solver(s); In some methods with different levels/fidelities. The section is called `models` in Multifidelity Monte Carlo to rreflect that several different baseline solvers can be combined here. |
+| `solver`/`models` | Setup of the baseline solver(s); In some methods with different levels/fidelities. The section is called `models` in Multifidelity Monte Carlo to reflect that several different baseline solvers can be combined here. |
 | `stoch_vars`      | A list of the used uncertain parameters including their distribution type and defining parameters |
 | `machine`         | Choice and setup of the machine that the run is performed on, such as a cluster or a local machine. Several machines can be listed and assigned to several parts of the simulation. |
 | `qois`            | A list of the quantities of interest (QoIs) which are evaluated. Several QoIs can be evaluated in one run. | 
 
-
+Single parameters are also explained in the Features chapter of this documentation.
 
 ## Tests
 
@@ -51,7 +51,7 @@ pytest ../src/tests.py
 ```
 The tests are all designed to run on a `local` machine, i.e. without a scheduler. The nine tests cover all combinations of three deterministic baseline solvers (`demonstrator_batch`, `demonstrator_single` and `internal`) and three UQ methods they use (`pce`, `mlmc` and `mfmc`). For details on machines, deterministic solvers and UQ methods, see the 'Features' section of this documentation. 
 
-The requirement for sucessful termination of each test is that the resulting mean and standard deviation equal reference values given in `src/tests.py` (To achieve reproducible results in the Monte Carlo simulations, the parameter `fixed_seed` in the parameter files is used to manually set the seed of the random number generator).
+The requirement for successful termination of each test is that the resulting mean and standard deviation equal reference values given in `src/tests.py` (To achieve reproducible results in the Monte Carlo simulations, the parameter `fixed_seed` in the parameter files is used to manually set the seed of the random number generator).
 
 Each test can also be run as an individual PoUnce run (with the syntax shown above). The according nine parameter files are located in the ini folder in the subdirectories `demonstrator_batch_local`, `demonstrator_single_local` and `internal_local`.
 
