+++ doc/200_features.md
@@ -1,36 +1,58 @@
-# Features
+# Features and detailed parameter description
 
+In this chapter, features of PoUnce are given and their configuration and use in the parameter files is explained. Please also refer to the explanations in the example parameter files in the `ini/demonstrator_batch_local` directory as well as `ini/ice/parameter_mfmc.yml`.
 
 ## UQ methods
 
-PoUnce currently features three UQ methods: 
+PoUnce currently features three UQ methods. They are configured via the `uq_method` and `sampling` sections of the parameter file: 
 
-- **The non-intrusive polynomial chaos method:** This method is implemented based on the ChaosPy [@ChaosPyFeinberg2015] package and includes the possibility of using sparse grids. Since this method in its basic form only requires one iteration of sample simulations of similar computational cost, the implementation is simple and can be easily extended to more advanced methods, such as adaptive sparse grids. 
-- **The multilevel Monte Carlo method:** If this method is run with two iterations (computing pilot samples in a first iteration and the rest of an approximately optimal number of samples in the second), it corresponds to the standard MLMC method as described in [@Giles2008] with a fixed set of resolution levels. However, in the implemented version, more iterations (usually three or four) can be used to carefully approach the optimal number of samples on every level while updating the estimates of these optimal numbers. This avoids overshooting the optimal sample numbers based on inaccurate estimates. Details on the method can be found in [@Beck2020]. 
-- **The multifidelity Monte Carlo method:** The method is implemented following [@Peherstorfer2016], with the additional option to re-use pilot samples for the eventual estimators.
+- **The non-intrusive polynomial chaos method (`_type: pce)`:** This method is implemented based on the ChaosPy [@ChaosPyFeinberg2015] package. The polynomial degree can be changed via `poly_deg` in the `sampleing` section of the parameter file. It includes the possibility of using sparse grids (`sparse_grid: true` in the `sampling` section). Since this method in its basic form only requires one iteration of sample simulations of similar computational cost, the implementation is simple and can be easily extended to more advanced methods, such as adaptive sparse grids. 
+- **The multilevel Monte Carlo method (`_type: mlmc)`:** If this method is run with two iterations (computing pilot samples in a first iteration and the rest of an approximately optimal number of samples in the second; achieved by setting `n_max_iter: 2` in the `uq_method` section), it corresponds to the standard MLMC method as described in [@Giles2008] with a fixed set of resolution levels. However, in the implemented version, more iterations (usually three or four) can be used to carefully approach the optimal number of samples on every level while updating the estimates of these optimal numbers. This avoids overshooting the optimal sample numbers based on inaccurate estimates. To calculate the number of samples in the intermediate iterations, a heuristic approach based on the current and the optimal number of samples can be used (`use_ci: False`), or a method based on confidence intervals described in Detail in [@Beck2020], (`use_ci: True`). Either the `total_work` or the estimated stochastic error (`eps`) can be prescribed the other is then optimized. If the MLMC method is used, further sections of the parameter file are affected: In `sampling`, `fixed_seed: True` can be used to achieve the same results with the random number generator in every run which makes results reproducible. The `solver` section contains the additional parameter `n_warmup_samples` which determines the number of samples on every level in the first iteration. It is given in this section so that different values can be specified for each level. One of the `qois` gets and additional parameter `optimized`, which determines which QoI is used to optimize the sample number on every level. 
[…]
