### 2023-01-15 #4683 @JakobBD
Hey @Nikoleta-v3 thank you for the review! I feel that your suggestions make a lot of sense and my changes have hopefully improved the documentation.

I have addressed them as follows:

> * In the ReadMe, in the `Quick start / basic run` section there is a typo.  It should be `python3 ../../src/pounce.py parameter_mlmc.yml`.

Thanks, I have fixed that.

> * Could you also include a sentence about the expected outcome? Tell the user that the results are printed in the terminal and also an output csv is exported.

Good point, I have added a sentence.

> * In the `Documentation` section could you please include a link to the documentation?

I have added the link.

> * In the manuscript line 55: it is meant to be interval or internal?

It is supposed to be internal. With that wording, I try to distinguish the scheduling of samples I do myself within a compute job from the cluster's job scheduling software which handles the queue of compute jobs of different users.

> 1. In the documentation the sections “Parameters” and “Features” should include more information. In the parameters you should walk the user over all the sections of a `yml` file (`general`, `uq_method`, …).  In the features you should include code snippets with examples of how different methods, solvers etc can be used. The example `yml` files are great as examples but this information should be more accessible/obvious to a user.

I have extended both sections as you suggested, expecially the "Features" one, which hopefully now explains all parameters and puts them in some context. Some of the details of the `uq_methods` parameters are beyond the scope of the documentation, but I am preparing a paper on this and will put a reference to it in the documentation once I have a draft.

> 2. To me pounce it’s the perfect project to be a command line tool. So I could just run `pounce parameter_mfmc.yml`. Is there a reason behind your decision to keep it in scripts format? 😊

I've added a `Makefile` which adds this possibility and I've explained its use in the documentation under  "Getting started / Command line tool"

> 3. You say that:
> > Output files will be located in the current working directory, so it is best to copy a parameter file to an empty directory and start pounce from there.
> Why not have the location and the name (if not given by the user) of the output file as arguments?

PoUnce runs produce various different files (such as for communication with the external baseline solver, restarting, and output), typically much more than the example from the "Getting started" section. So it makes sense to all have them in a separate directory without other runs to keep an overview. Manually creating that directory and starting PoUnce from it doesn't seem like a major inconvenience to me. This also makes the use of the restart capability clearer, as `pounce -r` looks for the `pounce.pickle` file in the current working directory. Otherwise, the same output directory flag would have to be given to PoUnce as when it was started for the first time. From an implementational point of view, file I/O happens at a lot of different places in the code (for example, each baseline solver has its own routines), so I'd have to find and change every single one to prepend the output directory to the file path, and I'd be afraid to miss one and introduce bugs. I hope that makes sense.
