### 2023-01-16 #4683 @Nikoleta-v3
@JakobBD initially, thank you very for addressing my comments 👍🏻

> It is supposed to be internal. With that wording, I try to distinguish the scheduling of samples I do myself within a compute job from the cluster's job scheduling software which handles the queue of compute jobs of different users.

I understand now; thank you for explaining.

> I have extended both sections as you suggested, expecially the "Features" one, which hopefully now explains all parameters and puts them in some context. Some of the details of the uq_methods parameters are beyond the scope of the documentation, but I am preparing a paper on this and will put a reference to it in the documentation once I have a draft.

This is great; thank you for working on the documentation! One minor issue, can you check that the documentation has built correctly? For example in [Workflow](https://github.com/JakobBD/pounce/blob/master/doc/400_workflow.md) I can see the `\textit{` commands 🤔

> PoUnce runs produce various different files (such as for communication with the external baseline solver, restarting, and output), typically much more than the example from the "Getting started" section. So it makes sense to all have them in a separate directory without other runs to keep an overview. Manually creating that directory and starting PoUnce from it doesn't seem like a major inconvenience to me. This also makes the use of the restart capability clearer, as pounce -r looks for the pounce.pickle file in the current working directory. Otherwise, the same output directory flag would have to be given to PoUnce as when it was started for the first time. From an implementational point of view, file I/O happens at a lot of different places in the code (for example, each baseline solver has its own routines), so I'd have to find and change every single one to prepend the output directory to the file path, and I'd be afraid to miss one and introduce bugs. I hope that makes sense.

Understood; thank you for your reply.

`pounce parameter_mlmc.yml` 🔥 👏🏻

I will wait for a few more days to see if the second reviewer has any further comments or thoughts and then we can move on with the submission 😄
