### 2023-01-18 #4683 @JakobBD
> This is great; thank you for working on the documentation! One minor issue, can you check that the documentation has built correctly? For example in [Workflow](https://github.com/JakobBD/pounce/blob/master/doc/400_workflow.md) I can see the `\textit{` commands 🤔

The markdown web interface is only a byproduct, primarily the documentation is built using pandoc (as stated in the [readme](https://github.com/JakobBD/pounce/tree/master/doc#readme)), that's why there are still some latex commands. It was chosen over plain markdown due to the better capabilities in pandoc to reference images and literature.

> I will wait for a few more days to see if the second reviewer has any further comments or thoughts and then we can move on with the submission 😄

Thanks! 🎉
