### 2023-02-01 #4683 @Nikoleta-v3
@JakobBD thank you for your patience!! We are good to go 👍🏻

At this point could you please:

-  (if necessary) Make a new tagged release of your software, and list the version tag of the archived version here. Or let me know if I should just use the release v1.0.0
 Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.

I can then move forward with accepting the submission 🎉
