### 2023-02-02 #4683 @JakobBD
Thanks!

- The version tag is [v1.0.0](https://github.com/JakobBD/pounce/releases/tag/v1.0.0)
- The Zenodo DOI is [10.5281/zenodo.7600634](https://doi.org/10.5281/zenodo.7600634)
