### 2023-02-03 #4683 @danielskatz
👋 @xuanxu - can you take a look at the proof, and specifically, the final reference. Note that the journal title starts with a ";", which is unusual, and there seems to be a space missing between the paper title and the journal title, perhaps because of the journal title starting with a special character?

It seems like this might be a bug somewhere in our pandoc/bibtex processing code - what do you think?
