### 2023-02-13 #4683 @tarleb
Thanks for the patience. I tried to debug this, but the only thing I can state for certain is that this is somehow due to our citation style (APA). Some styles like *chicago-author-date* do not have this problem, while the *nature* style keeps the period but drops the space.

I'll will keep digging, but for now I'd recommend this workaround:

```
  journal = {​{\tt ;login:} The USENIX Magazine},
```

This is a hack which does two things: First, the initial character in the journal name above is a ZERO WIDTH SPACE. This ensures that the journal name does not start with punctuation, so the article title gets properly terminated with a period. Second, we mark the string `;login:` as code, which prevents unwanted capitalization to *;Login:*; this seems in line with how the journal title is styled on the cover.
