### 2023-02-14 438ab4e
workaround for pandoc reference formatting bug
