### 2023-02-15 #4683 @tarleb
The hack wasn't applied in full, the ZERO WIDTH SPACE is missing. I'll open a PR.
