### 2023-02-15 a1bb8bb
Paper: Add zero width space before journal title starting with `;`
