+++ paper/paper.bib
@@ -183,7 +183,7 @@
   doi = {10.5281/zenodo.16303},
   interhash = {b829bb54862d2644509ad3d5fa2a2e09},
   intrahash = {ab8ea12a0dffc7a4f4d5e972c8fb4ad8},
-  journal = {{\tt ;login:} The USENIX Magazine},
+  journal = {​{\tt ;login:} The USENIX Magazine},
   keywords = {imported},
   month = Feb,
   number = 1,
